requirements.txt
python-dotenv
snowflake
snowflake-snowpark-python
pydantic
